FOR-LOOP COMPLETE
I finally understood the logic behind the For-loop so I am able to execute it the way that I wanted it to. I was only able to execute one line inside of the FOR-loop but now I can have multiple lines in my FOR-loop and have them working properly as intended.