JUMP
I need to focus on this subject called a "JUMP" because I am keeping on printing the same values inside of my FOR-loop although I already went through all of the loop. This may have been an issue with having created a for each loop since we need to pass through each instruction as opposed to "JUMPING" to other instructions we care about.